+++ Example/Shared/TwitterView.swift
@@ -9,7 +9,7 @@ import SwiftUI
 import PagerTabStripView
 
 struct TwitterView: View {
-    private var swipeGestureEnabled: Bool
+    @State var swipeGestureEnabled: Bool
     @State var selection = 2
 
     @ObservedObject var firstModel = TweetsModel()
@@ -23,35 +23,36 @@ struct TwitterView: View {
         self.swipeGestureEnabled = swipeGestureEnabled
     }
 
-    var body: some View {
-        PagerTabStripView(swipeGestureEnabled: swipeGestureEnabled, selection: $selection) {
-            PostsList(isLoading: $firstModel.isLoading, items: firstModel.posts).pagerTabItem {
-                TwitterNavBarItem(title: "First big width")
-            }
-
-            PostsList(isLoading: $secondModel.isLoading, items: secondModel.posts).pagerTabItem {
-                TwitterNavBarItem(title: "Short")
-            }
-
-            PostsList(isLoading: $thirdModel.isLoading, items: thirdModel.posts, withDescription: false).pagerTabItem {
-                TwitterNavBarItem(title: "Medium width")
-            }
-
-            PostsList(isLoading: $fourthModel.isLoading, items: fourthModel.posts).pagerTabItem {
-                TwitterNavBarItem(title: "Second big width")
-            }
-
-            PostsList(isLoading: $fifthModel.isLoading, items: fifthModel.posts, withDescription: false).pagerTabItem {
-                TwitterNavBarItem(title: "Medium width")
-            }
-
-            PostsList(isLoading: $sixthModel.isLoading, items: sixthModel.posts).pagerTabItem {
-                TwitterNavBarItem(title: "Mini")
-            }
+    @MainActor var body: some View {
+        PagerTabStripView(swipeGestureEnabled: $swipeGestureEnabled, selection: $selection) {
+            PostsList(isLoading: $firstModel.isLoading, items: firstModel.posts)
+                .pagerTabItem {
+                    TwitterNavBarItem(title: "First big width")
+                }
+            PostsList(isLoading: $secondModel.isLoading, items: secondModel.posts)
+                .pagerTabItem {
+                    TwitterNavBarItem(title: "Short")
+                }
+            PostsList(isLoading: $thirdModel.isLoading, items: thirdModel.posts, withDescription: false)
+                .pagerTabItem {
+                    TwitterNavBarItem(title: "Medium width")
+                }
+            PostsList(isLoading: $fourthModel.isLoading, items: fourthModel.posts)
+                .pagerTabItem {
+                    TwitterNavBarItem(title: "Second big width")
+                }
+            PostsList(isLoading: $fifthModel.isLoading, items: fifthModel.posts, withDescription: false)
+                .pagerTabItem {
+                    TwitterNavBarItem(title: "Medium width")
+                }
+            PostsList(isLoading: $sixthModel.isLoading, items: sixthModel.posts)
+                .pagerTabItem {
+                    TwitterNavBarItem(title: "Mini")
+                }
         }
-        .frame(alignment: .center)
-        .pagerTabStripViewStyle(.scrollableBarButton(indicatorBarColor: .blue, tabItemSpacing: 15, tabItemHeight: 50))
-        .navigationBarTitleDisplayMode(.inline)
+        .pagerTabStripViewStyle(.scrollableBarButton(placedInToolbar: false, tabItemSpacing: 15, tabItemHeight: 40, indicatorView: {
+            Rectangle().fill(.blue).padding(EdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 0))
+        }))
     }
 }