+++ Sources/NavigationBarStyles/IndicatorBarView.swift
@@ -8,26 +8,25 @@
 import Foundation
 import SwiftUI
 
-internal struct IndicatorBarView<Indicator: View>: View {
+internal struct IndicatorBarView<Indicator>: View where Indicator: View {
     @EnvironmentObject private var dataStore: DataStore
-
-    @ViewBuilder var indicator: Indicator
+    @ViewBuilder var indicator: () -> Indicator
 
     var body: some View {
-        HStack {
-            let totalItemWidth = (settings.width - (style.tabItemSpacing * CGFloat(dataStore.itemsCount - 1)))
-            let navBarItemWidth = totalItemWidth / CGFloat(dataStore.itemsCount)
-            if let width = navBarItemWidth, width > 0, width <= settings.width {
-                let x = -settings.contentOffset / CGFloat(dataStore.itemsCount) + width / 2
-
-                indicator
-                    .foregroundColor(style.indicatorBarColor)
-                    .animation(.default)
-                    .frame(width: width)
-                    .position(x: x, y: 0)
+        if let internalStyle = style as? PagerWithIndicatorStyle {
+            HStack {
+                let totalItemWidth = (settings.width - (internalStyle.tabItemSpacing * CGFloat(dataStore.itemsCount - 1)))
+                let navBarItemWidth = totalItemWidth / CGFloat(dataStore.itemsCount)
+                if let navBarItemWidth, navBarItemWidth > 0, navBarItemWidth <= settings.width {
+                    let x = -settings.contentOffset / CGFloat(dataStore.itemsCount) + navBarItemWidth / 2
+                    indicator()
+                        .animation(.default)
+                        .frame(width: navBarItemWidth)
+                        .position(x: x, y: internalStyle.indicatorViewHeight / 2)
+                }
             }
+            .frame(height: internalStyle.indicatorViewHeight)
         }
-        .frame(height: style.indicatorBarHeight)
     }
 
     @Environment(\.pagerStyle) var style: PagerStyle