+++ Sources/PagerTabItemModifier.swift
@@ -15,18 +15,16 @@ struct PagerTabItemModifier<NavTabView: View>: ViewModifier {
         self.navTabView = navTabView
     }
 
-    func body(content: Content) -> some View {
+    @MainActor func body(content: Content) -> some View {
         GeometryReader { reader in
             content
                 .onAppear {
-                    DispatchQueue.main.async {
-                        let frame = reader.frame(in: .named("PagerViewScrollView"))
-                        index = Int(round(frame.minX / frame.width))
-                        let tabView = navTabView()
-                        let tabViewDelegate = tabView as? PagerTabViewDelegate
-                        dataStore.setView(AnyView(tabView), at: index)
-                        dataStore.setTabViewDelegate(tabViewDelegate, at: index)
-                    }
+                    let frame = reader.frame(in: .named("PagerViewScrollView"))
+                    index = Int(round(frame.minX / frame.width))
+                    let tabView = navTabView()
+                    let tabViewDelegate = tabView as? PagerTabViewDelegate
+                    dataStore.setView(AnyView(tabView), at: index)
+                    dataStore.setTabViewDelegate(tabViewDelegate, at: index)
                 }.onDisappear {
                     dataStore.items[index]?.tabViewDelegate?.setState(state: .normal)
                     dataStore.remove(at: index)