+++ Sources/PagerTabStripView.swift
@@ -14,32 +14,20 @@ class PagerSettings: ObservableObject {
 @available(iOS 14.0, *)
 public struct PagerTabStripView<Content>: View where Content: View {
     private var content: () -> Content
-
-    @Binding private var selectionBiding: Int
+    private var swipeGestureEnabled: Binding<Bool>
+    private var selection: Binding<Int>?
     @State private var selectionState = 0
     @StateObject private var settings: PagerSettings
-    private var useBinding: Bool
-    private let swipeGestureEnabled: Bool
 
-    public init(swipeGestureEnabled: Bool = true,
-                selection: Binding<Int>? = nil,
-                @ViewBuilder content: @escaping () -> Content) {
-        self.content = content
-        if let selection = selection {
-            useBinding = true
-            self._selectionBiding = selection
-        } else {
-            useBinding = false
-            self._selectionBiding = .constant(0)
-        }
+    public init(swipeGestureEnabled: Binding<Bool> = .constant(true), selection: Binding<Int>? = nil, @ViewBuilder content: @escaping () -> Content) {
         self.swipeGestureEnabled = swipeGestureEnabled
+        self.selection = selection
+        self.content = content
         self._settings = StateObject(wrappedValue: PagerSettings())
     }
 
-    public var body: some View {
-        WrapperPagerTabStripView(swipeGestureEnabled: swipeGestureEnabled,
-                                 selection: useBinding ? $selectionBiding : $selectionState,
-                                 content: content)
+    @MainActor public var body: some View {
+        WrapperPagerTabStripView(swipeGestureEnabled: swipeGestureEnabled, selection: selection ?? $selectionState, content: content)
             .environmentObject(self.settings)
     }
 }
@@ -49,34 +37,26 @@ private struct WrapperPagerTabStripView<Content>: View where Content: View {
     private var content: () -> Content
 
     @StateObject private var dataStore = DataStore()
-
     @Environment(\.pagerStyle) var style: PagerStyle
     @EnvironmentObject private var settings: PagerSettings
     @Binding var selection: Int {
         didSet {
-            if selection < 0 {
-                selection = oldValue
-            }
+            if selection < 0 { selection = oldValue }
         }
     }
     @State private var currentOffset: CGFloat = 0 {
-        didSet {
-            self.settings.contentOffset = currentOffset
-        }
+        didSet { settings.contentOffset = currentOffset }
     }
     @GestureState private var translation: CGFloat = 0
+    @Binding private var swipeGestureEnabled: Bool
 
-    private let swipeGestureEnabled: Bool
-
-    public init(swipeGestureEnabled: Bool = true,
-                selection: Binding<Int>,
-                @ViewBuilder content: @escaping () -> Content) {
-        self.swipeGestureEnabled = swipeGestureEnabled
-        self.content = content
+    public init(swipeGestureEnabled: Binding<Bool>, selection: Binding<Int>, @ViewBuilder content: @escaping () -> Content) {
+        self._swipeGestureEnabled = swipeGestureEnabled
         self._selection = selection
+        self.content = content
     }
 
-    public var body: some View {
+    @MainActor public var body: some View {
         GeometryReader { gproxy in
             LazyHStack(spacing: 0) {
                 content()
@@ -87,9 +67,8 @@ private struct WrapperPagerTabStripView<Content>: View where Content: View {
             .offset(x: self.translation)
             .animation(style.pagerAnimation, value: selection)
             .animation(.interactiveSpring(response: 0.15, dampingFraction: 0.86, blendDuration: 0.25), value: translation)
-            .gesture(
+            .gesture(!swipeGestureEnabled ? nil :
                 DragGesture(minimumDistance: 25).updating(self.$translation) { value, state, _ in
-                    guard swipeGestureEnabled else { return }
                     if selection == 0 && value.translation.width > 0 {
                         let valueWidth = value.translation.width
                         let normTrans = valueWidth / (gproxy.size.width + 50)
@@ -104,7 +83,6 @@ private struct WrapperPagerTabStripView<Content>: View where Content: View {
                         state = value.translation.width
                     }
                 }.onEnded { value in
-                    guard swipeGestureEnabled else { return }
                     let offset = value.predictedEndTranslation.width / gproxy.size.width
                     let newPredictedIndex = (CGFloat(selection) - offset).rounded()
                     let newIndex = min(max(Int(newPredictedIndex), 0), dataStore.itemsCount - 1)
@@ -120,15 +98,15 @@ private struct WrapperPagerTabStripView<Content>: View where Content: View {
             )
             .onAppear(perform: {
                 let geo = gproxy.frame(in: .local)
-                self.settings.width = geo.width
-                self.currentOffset = -(CGFloat(selection) * geo.width)
+                settings.width = geo.width
+                currentOffset = -(CGFloat(selection) * geo.width)
             })
             .onChange(of: gproxy.frame(in: .local), perform: { geo in
-                self.settings.width = geo.width
-                self.currentOffset = -(CGFloat(selection) * geo.width)
+                settings.width = geo.width
+                currentOffset = -(CGFloat(selection) * geo.width)
             })
             .onChange(of: self.selection) { [selection] newIndex in
-                self.currentOffset = -(CGFloat(newIndex) * gproxy.size.width)
+                currentOffset = -(CGFloat(newIndex) * gproxy.size.width)
                 dataStore.items[newIndex]?.appearCallback?()
                 dataStore.items[selection]?.tabViewDelegate?.setState(state: .normal)
                 if let tabViewDelegate = dataStore.items[newIndex]?.tabViewDelegate, newIndex != selection {
@@ -136,16 +114,16 @@ private struct WrapperPagerTabStripView<Content>: View where Content: View {
                 }
             }
             .onChange(of: translation) { _ in
-                self.settings.contentOffset = translation - CGFloat(selection)*gproxy.size.width
+                settings.contentOffset = translation - CGFloat(selection)*gproxy.size.width
             }
             .onChange(of: dataStore.itemsCount) { _ in
-                self.selection = selection >= dataStore.itemsCount ? dataStore.itemsCount - 1 : selection
+                selection = selection >= dataStore.itemsCount ? dataStore.itemsCount - 1 : selection
                 dataStore.items[selection]?.tabViewDelegate?.setState(state: .selected)
                 dataStore.items[selection]?.appearCallback?()
             }
         }
         .modifier(NavBarModifier(selection: $selection))
-        .environmentObject(self.dataStore)
+        .environmentObject(dataStore)
         .clipped()
     }